Splitting the util module
