+++ core/src/main/java/ru/funpay4j/client/jsoup/JsoupFunPayParser.java
@@ -17,10 +17,10 @@ package ru.funpay4j.client.jsoup;
 import com.google.gson.JsonParser;
 import lombok.NonNull;
 import okhttp3.*;
-import org.jetbrains.annotations.Nullable;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
+import ru.funpay4j.utils.FunPayUserUtil;
 import ru.funpay4j.client.FunPayParser;
 import ru.funpay4j.core.exceptions.FunPayApiException;
 import ru.funpay4j.core.exceptions.offer.OfferNotFoundException;
@@ -34,7 +34,6 @@ import ru.funpay4j.core.objects.lot.LotCounter;
 import ru.funpay4j.core.objects.offer.Offer;
 import ru.funpay4j.core.objects.offer.PreviewOffer;
 import ru.funpay4j.core.objects.user.*;
-import ru.funpay4j.util.FunPayUserUtil;
 
 import java.io.IOException;
 import java.text.ParseException;
